+++ meridian/analysis/visualizer.py
@@ -1638,6 +1638,101 @@ class MediaSummary:
     self._marginal_roi_by_reach = marginal_roi_by_reach
     self._non_media_baseline_values = non_media_baseline_values
 
+  def plot_channel_contribution_area_chart(self) -> alt.Chart:
+    """Plots a stacked area chart of the contribution share per channel by time.
+
+    Returns:
+      An Altair plot showing the contribution share per channel by time.
+    """
+    outcome_df = self._transform_contribution_metrics(
+        include_non_paid=True, aggregate_times=False
+    )
+    outcome_df[c.CHANNEL] = outcome_df[c.CHANNEL].str.upper()
+
+    # Ensure proper ordering for the stacked area chart. Baseline should be at
+    # the bottom.  Separate the *stacking* order from the *legend* order.
+    stack_order = sorted([
+        channel
+        for channel in outcome_df[c.CHANNEL].unique()
+        if channel != c.BASELINE.upper()
+    ]) + [c.BASELINE.upper()]
+
+    legend_order = [c.BASELINE.upper()] + sorted([
+        channel
+        for channel in outcome_df[c.CHANNEL].unique()
+        if channel != c.BASELINE.upper()
+    ])
+
+    # Get the minimum incremental outcome for baseline across all time periods
+    # as the lower bound for the stacked area chart.
+    min_y = (
+        outcome_df[outcome_df[c.CHANNEL] == c.BASELINE.upper()]
+        .groupby(c.TIME)[c.INCREMENTAL_OUTCOME]
+        .min()
+        .min()
+    )
+
+    plot = (
+        alt.Chart(outcome_df, width=c.VEGALITE_FACET_LARGE_WIDTH)
+        .mark_area()
+        .transform_calculate(
+            sort_channel=f'indexof({stack_order}, datum.channel)'
+        )
+        .encode(
+            x=alt.X(
+                f'{c.TIME}:T',
+                title='Time period',
+                axis=alt.Axis(
+                    format='%Y Q%q',
+                    grid=False,
+                    tickCount=8,
+                    domainColor=c.GREY_300,
+                ),
+            ),
+            y=alt.Y(
+                f'{c.INCREMENTAL_OUTCOME}:Q',
+                title=(
+                    c.REVENUE.title()
+                    if self._meridian.input_data.revenue_per_kpi is not None
+                    else c.KPI.upper()
+                ),
+                axis=alt.Axis(
+                    ticks=False,
+                    domain=False,
+                    tickCount=5,
+                    labelPadding=c.PADDING_10,
+                    labelExpr=formatter.compact_number_expr(),
+                    **formatter.Y_AXIS_TITLE_CONFIG,
+                ),
+                scale=alt.Scale(domainMin=min_y, clamp=True),
+            ),
+            color=alt.Color(
+                f'{c.CHANNEL}:N',
+                legend=alt.Legend(
+                    labelFontSize=c.AXIS_FONT_SIZE,
+                    labelFont=c.FONT_ROBOTO,
+                    title=None,
+                ),
+                scale=alt.Scale(domain=legend_order),
+                sort=legend_order,
+            ),
+            tooltip=[
+                alt.Tooltip(f'{c.TIME}:T', format='%Y-%m-%d'),
+                c.CHANNEL,
+                alt.Tooltip(f'{c.INCREMENTAL_OUTCOME}:Q', format=',.2f'),
+            ],
+            order=alt.Order('sort_channel:N', sort='descending'),
+        )
+        .properties(
+            title=formatter.custom_title_params(
+                summary_text.CHANNEL_CONTRIB_BY_TIME_CHART_TITLE
+            ),
+        )
+        .configure_axis(titlePadding=c.PADDING_10, **formatter.TEXT_CONFIG)
+        .configure_view(strokeOpacity=0)
+    )
+    return plot
+
   def plot_contribution_waterfall_chart(self) -> alt.Chart:
     """Plots a waterfall chart of the contribution share per channel.
 
@@ -1661,7 +1756,7 @@ class MediaSummary:
     num_channels = len(outcome_df[c.CHANNEL])
 
     base = (
-        alt.Chart(outcome_df)
+        alt.Chart(outcome_df, width=c.VEGALITE_FACET_LARGE_WIDTH)
         .transform_window(
             sum_outcome=f'sum({c.PCT_OF_CONTRIBUTION})',
             kwargs=f'lead({c.CHANNEL})',
@@ -1722,7 +1817,6 @@ class MediaSummary:
             ),
             height=c.BAR_SIZE * num_channels
             + c.BAR_SIZE * 2 * c.SCALED_PADDING,
-            width=500,
         )
         .configure_axis(titlePadding=c.PADDING_10, **formatter.TEXT_CONFIG)
         .configure_view(strokeOpacity=0)